Document when '--process-dependency-links' has to be used
When installing from source or preparing a contributor development
environment, use '--process-dependency-links' pip argument to be able to
install the latest 'master' version of Resolwe from git.
Specify an exact version of Resolwe (during development phase of Resolwe
Bioinformatics) and change it to 'master' so that it will be always
installed from git.