Fix Mapping search for source_id / target_id
If one did a Mappings query with multiple source_id/target_id as
query parameter the response would be empty or Error. There are
multiple options for one to get multiple Mappings from api. These
are the actual query parameters in request URL:

1. source_id="ID1,ID2"
2. source_id=["ID1", "ID2"]
3. source_id__in="ID1,ID2"
4. source_id__in=["ID1", "ID2"]

The main point is that value of query parameter must be casted to
list as some point. Since source_id is a custom filter the "__in"
snytax is not allowed (3. and 4. are not an option) but there is
no place in code where 1. and 2. would be casted to list. Therefore
the result of 1. or 2 was empty and 3. or 4. raised error.

This solution removes source_id (and targed_id) as custom filters
and places them as normal filtering parameters. Therefore "__in"
syntax is supported and the value of query parameter is splitted by
commma and casted to list. This is also sonsistent with how Feature
filter works for feature_id.

The regression of this solution is is the size of the list that can
be handled by ES in such cases. It reduces to 512 (not exact, but
1024 fails) by unknown reasons(ideas anyone?). However, in practice
this means that before one could filter by max 1 source_id and now
the limit is 512, which seems way better.